Adjusted styles ; Added add to cart functionality

+++ applications/models.py
@@ -67,6 +67,6 @@ class Cart(db.Model):
     id = db.Column(db.Integer , primary_key = True , autoincrement = True)
     user_id = db.Column(db.Integer, db.ForeignKey('user.id') , nullable = False)
     product_id = db.Column(db.Integer , db.ForeignKey('product.id') , nullable = False)
-    quantity = db.Column(db.Integer , nullable = False)
-    price = db.Column(db.Integer , nullable = False)
+    quantity = db.Column(db.Float, nullable = False)
+    price = db.Column(db.Float , nullable = False)
     date = db.Column(db.Date , nullable = False)
